fix iteration 2 and export product
