+++ src/components/ManageGoods/ChooseTypeDropdown.tsx
@@ -6,7 +6,7 @@ import AddPlusIcon from "../icons/AddPlusIcon"
 import ArrowDownIcon from "../icons/ArrowDownIcon"
 import SearchIcon from "../icons/SearchIcon"
 import AddSupplierPopup from "./AddSupplierPopup"
-import AddTypePopup from "./AddTypePopup"
+import AddTypePopup from "./ManageTypeGoods/AddTypePopup"
 import AddTypePopupInProductAdd from "./AddTypePopupInProductAdd"
 
 function ChooseTypeDropdown({
@@ -108,7 +108,6 @@ function ChooseTypeDropdown({
           className=""
           placeholder="Tìm kiếm hoặc nhập mới"
         />
-        <AddTypePopupInProductAdd />
         <div
           id="list-dropdown"
           className="smooth-transform z-50 flex w-full flex-col gap-1 bg-[#fff] pb-3 max-h-[250px] overflow-y-auto"
