+++ src/components/ManageGoods/ManageSuppliers/AddSupplier.tsx
@@ -1,22 +1,22 @@
 import React, { useEffect, useState } from "react"
-import PrimaryInput from "../PrimaryInput"
-import PrimaryTextArea from "../PrimaryTextArea"
-import SmallTitle from "../SmallTitle"
-import PrimaryBtn from "../PrimaryBtn"
+import PrimaryInput from "../../PrimaryInput"
+import PrimaryTextArea from "../../PrimaryTextArea"
+import SmallTitle from "../../SmallTitle"
+import PrimaryBtn from "../../PrimaryBtn"
 import { useMutation, useQueries } from "react-query"
 import { toast } from "react-toastify"
 import { useRouter } from "next/router"
-import { addNewProduct } from "../../apis/product-module"
-import CityDropDown from "../CityDropDown"
-import WardDropDown from "../WardDropDown"
-import DistrictDropDown from "../DistrictDropDown"
+import { addNewProduct } from "../../../apis/product-module"
+import CityDropDown from "../../CityDropDown"
+import WardDropDown from "../../WardDropDown"
+import DistrictDropDown from "../../DistrictDropDown"
 import {
   getListCity,
   getListDistrictByCode,
   getListWardByCode,
-} from "../../apis/search-country-module"
-import { addNewSupplier } from "../../apis/supplier-module"
-import ConfirmPopup from "../ConfirmPopup"
+} from "../../../apis/search-country-module"
+import { addNewSupplier } from "../../../apis/supplier-module"
+import ConfirmPopup from "../../ConfirmPopup"
 
 interface Supplier {
   supplierId: number
@@ -216,7 +216,7 @@ function AddSupplier(props) {
               setSupplier({ ...supplier, note: e.target.value })
             }}
           />
-          <div className="flex items-center absolute-right">
+          <div className="flex items-center absolute-right mt-6">
             <div className="flex flex-col gap-4">
               <div className="grid items-center justify-between fle w-full gap-4 md:grid-cols-2 ">
                 <ConfirmPopup
