+++ src/components/ManageGoods/ManageSuppliers/ManageSuppliers.tsx
@@ -1,24 +1,24 @@
 import React, { useEffect, useState } from "react"
-import DemoDropDown from "../DemoDropDown"
-import DownloadIcon from "../icons/DownloadIcon"
-import PlusIcon from "../icons/PlusIcon"
-import UploadIcon from "../icons/UploadIcon"
-import PrimaryBtn from "../PrimaryBtn"
-import SearchInput from "../SearchInput"
-import ShowLabelBar from "../Filter/ShowLabelBar"
-import Table from "../Table"
-import Pagination from "../Pagination"
+import DemoDropDown from "../../DemoDropDown"
+import DownloadIcon from "../../icons/DownloadIcon"
+import PlusIcon from "../../icons/PlusIcon"
+import UploadIcon from "../../icons/UploadIcon"
+import PrimaryBtn from "../../PrimaryBtn"
+import SearchInput from "../../SearchInput"
+import ShowLabelBar from "../../Filter/ShowLabelBar"
+import Table from "../../Table"
+import Pagination from "../../Pagination"
 import Link from "next/link"
-import ShowDetailIcon from "../icons/ShowDetailIcon"
+import ShowDetailIcon from "../../icons/ShowDetailIcon"
 import BigNumber from "bignumber.js"
-import useDebounce from "../../hooks/useDebounce"
+import useDebounce from "../../../hooks/useDebounce"
 import { useQueries } from "react-query"
 import {
   getListSupplier,
   getListExportSupplier,
-} from "../../apis/supplier-module"
+} from "../../../apis/supplier-module"
 import * as XLSX from "xlsx/xlsx"
-import EditIcon from "../icons/EditIcon"
+import EditIcon from "../../icons/EditIcon"
 import { format, parseISO } from "date-fns"
 
 const columns = [
