+++ src/components/ManageGoods/ManageSuppliers/SupplierDetail.tsx
@@ -1,20 +1,20 @@
 import React, { useEffect, useState } from "react"
-import SmallTitle from "../SmallTitle"
-import DemoDropDown from "../DemoDropDown"
-import PrimaryBtn from "../PrimaryBtn"
+import SmallTitle from "../../SmallTitle"
+import DemoDropDown from "../../DemoDropDown"
+import PrimaryBtn from "../../PrimaryBtn"
 import { useMutation, useQueries } from "react-query"
 import { toast } from "react-toastify"
 import { useRouter } from "next/router"
-import { addNewProduct, getListProduct } from "../../apis/product-module"
-import SearchInput from "../SearchInput"
-import Table from "../Table"
+import { addNewProduct, getListProduct } from "../../../apis/product-module"
+import SearchInput from "../../SearchInput"
+import Table from "../../Table"
 import { format, parseISO } from "date-fns"
 import Link from "next/link"
-import ShowDetailIcon from "../icons/ShowDetailIcon"
+import ShowDetailIcon from "../../icons/ShowDetailIcon"
 import BigNumber from "bignumber.js"
-import { getSupplierDetail } from "../../apis/supplier-module"
-import Pagination from "../Pagination"
-import useDebounce from "../../hooks/useDebounce"
+import { getSupplierDetail } from "../../../apis/supplier-module"
+import Pagination from "../../Pagination"
+import useDebounce from "../../../hooks/useDebounce"
 
 const columns = [
   {
