+++ src/constants/APIConfig.tsx
@@ -1,9 +1,9 @@
 // export const baseUrl = process.env.NEXT_PUBLIC_BASE_URL
 
 // Dev baseUrl
-export const baseUrl = `https://imsdservice.azurewebsites.net/api`
+// export const baseUrl = `https://imsdservice.azurewebsites.net/api`
 
-// export const baseUrl = `https://localhost:7265/api`
+export const baseUrl = `https://localhost:7265/api`
 
 // City, district, ward
 export const countryUrl = `https://provinces.open-api.vn/api`
