+++ src/pages/supplier-detail/[supplierid].tsx
@@ -1,6 +1,6 @@
 import React from "react"
 import Layout from "../../components/Layout"
-import SupplierDetail from "../../components/ManageSuppliers/SupplierDetail"
+import SupplierDetail from "../../components/ManageGoods/ManageSuppliers/SupplierDetail"
 
 function detailSupplier(props) {
   return (
